delete and move classes
