+++ app/src/main/java/io/github/sds100/keymapper/actions/DisplayActionUseCase.kt
@@ -1,14 +1,17 @@
-package io.github.sds100.keymapper.mappings
+package io.github.sds100.keymapper.actions
 
 import android.graphics.drawable.Drawable
-import io.github.sds100.keymapper.constraints.GetConstraintErrorUseCase
 import io.github.sds100.keymapper.util.Error
 import io.github.sds100.keymapper.util.Result
+import kotlinx.coroutines.flow.Flow
 
-interface DisplayConstraintUseCase : GetConstraintErrorUseCase {
+interface DisplayActionUseCase : GetActionErrorUseCase {
+    val showDeviceDescriptors: Flow<Boolean>
     fun getAppName(packageName: String): Result<String>
     fun getAppIcon(packageName: String): Result<Drawable>
     fun getInputMethodLabel(imeId: String): Result<String>
-    fun neverShowDndTriggerError()
     suspend fun fixError(error: Error)
+    fun neverShowDndTriggerError()
+    fun startAccessibilityService(): Boolean
+    fun restartAccessibilityService(): Boolean
 }
