+++ app/src/main/java/io/github/sds100/keymapper/sorting/comparators/KeyMapActionsComparator.kt
@@ -1,7 +1,7 @@
 package io.github.sds100.keymapper.sorting.comparators
 
 import io.github.sds100.keymapper.actions.ActionData
-import io.github.sds100.keymapper.mappings.DisplayActionUseCase
+import io.github.sds100.keymapper.actions.DisplayActionUseCase
 import io.github.sds100.keymapper.mappings.keymaps.KeyMap
 import io.github.sds100.keymapper.util.Result
 import io.github.sds100.keymapper.util.Success
